Added full Lyceum suite (without Trajectories for now)
