+++ ext/VisualiserExt/visualiser.jl
@@ -1,9 +1,6 @@
-# Many elements of this file are taken from https://github.com/Lyceum/LyceumMuJoCoViz.jl
-# It is a work-in-progress
+# Most this file is slightly modified from https://github.com/Lyceum/LyceumMuJoCoViz.jl
 
-# TODO: Move all of this into the main VisualiserExt.jl script so it's all together
-
-import Base.RefValue
+import Base: @lock, @lock_nofail
 
 if isdefined(Base, :get_extension)
     import MuJoCo.LibMuJoCo: Model, Data
@@ -24,6 +21,7 @@ end
 const Maybe{T} = Union{T, Nothing}  # From LyceumBase.jl
 const MAXGEOM = 10000               # preallocated geom array in mjvScene
 const MIN_REFRESHRATE = 30          # minimum rate when sim can't run at native refresh rate
+const RNDGAMMA = 0.9
 
 const RES_HD = (1280, 720)
 const RES_FHD = (1920, 1080)
@@ -41,14 +39,106 @@ include("defaulthandlers.jl")
 
 # ----------------------------------------------------------------------------------
 
-# TODO: Remove this later
-function MuJoCoViewer(m::Model, d::Data; controller=nothing)
+"""
+Add docstring here
+"""
+function visualise(m::Model, d::Data; controller=nothing)
     modes = EngineMode[PassiveDynamics()]
     !isnothing(controller) && push!(modes, Controller(controller))
-    return Engine(default_windowsize(), m, d, Tuple(modes))
+    e = Engine(default_windowsize(), m, d, Tuple(modes))
+    run(e)
 end
 
-# From LyceumMuJoCoViz.jl
+"""
+Run the visualiser engine
+"""
+function run(e::Engine)
+
+    # Render the first frame before opening window
+    prepare!(e)
+    e.ui.refreshrate = GetRefreshRate()
+    e.ui.lastrender = time()
+    GLFW.ShowWindow(e.manager.state.window)
+
+    # Run the simulation/mode in a different thread
+    modetask = Threads.@spawn runphysics(e)
+
+    # Print help info
+    println(ASCII)
+    println("Press \"F1\" to show the help message.")
+
+    # Run the visuals
+    runui(e)
+    wait(modetask)
+    return nothing
+end
+
+"""
+Run the UI
+"""
+function runui(e::Engine)
+    shouldexit = false
+    trecord = 0.0
+    try
+        while !shouldexit
+
+            # Check for window interaction and prepare visualisation
+            @lock e.phys.lock begin
+                GLFW.PollEvents()
+                prepare!(e)
+            end
+
+            # Render
+            render(e)
+            trender = time()
+
+            # Match the rendering rate to desired rates
+            rt = 1 / (trender - e.ui.lastrender)
+            @lock e.ui.lock begin
+                e.ui.refreshrate = RNDGAMMA * e.ui.refreshrate + (1 - RNDGAMMA) * rt
+                e.ui.lastrender = trender
+                shouldexit = e.ui.shouldexit | GLFW.WindowShouldClose(e.manager.state.window)
+            end
+
+            # Handle frame recording
+            tnow = time()
+            if e.ffmpeghandle !== nothing && tnow - trecord > 1 / e.min_refreshrate
+                trecord = tnow
+                recordframe(e)
+            end
+            yield() # Visualisation should give way to running the physics model
+        end
+    finally
+        @lock e.ui.lock begin
+            e.ui.shouldexit = true
+        end
+        GLFW.DestroyWindow(e.manager.state.window)
+    end
+    return
+end
+
+"""
+Prepare the visualisation engine for rendering
+"""
+function prepare!(e::Engine)
+    ui, p = e.ui, e.phys
+    m, d = p.model, p.data
+    LibMuJoCo.mjv_updateScene(
+        m.internal_pointer, 
+        d.internal_pointer, 
+        ui.vopt.internal_pointer, 
+        p.pert.internal_pointer, 
+        ui.cam.internal_pointer, 
+        LibMuJoCo.mjCAT_ALL,
+        ui.scn.internal_pointer
+    )
+    prepare!(ui, p, mode(e))
+    return e
+end
+
+"""
+Render a frame
+"""
 function render(e::Engine)
     w, h = GLFW.GetFramebufferSize(e.manager.state.window)
     rect = mjrRect(Cint(0), Cint(0), Cint(w), Cint(h))
@@ -59,35 +149,48 @@ function render(e::Engine)
 end
 
 """
-    render!(e::Engine, m::Model, d::Data)
+Run the MuJoCo model.
 
-Render the viewer given current model and data.
-
-# TODO: Add all extra stuff equivalent to `runui(e::Engine)` in `LyceumMuJoCoViz.jl` file
+This function handles simulating the model in pause, forward, and reverse mode.
+Note that reverse mode is only implemented for the `Trajectory` EngineMode.
 """
-function render!(e::Engine, m::Model, d::Data)
-    LibMuJoCo.mjv_updateScene(
-        m.internal_pointer, 
-        d.internal_pointer, 
-        e.ui.vopt.internal_pointer, 
-        C_NULL, 
-        e.ui.cam.internal_pointer, 
-        LibMuJoCo.mjCAT_ALL, 
-        e.ui.scn.internal_pointer
-    )
-    render(e)
-    GLFW.PollEvents()
-    e.should_close = e.ui.shouldexit | GLFW.WindowShouldClose(e.manager.state.window)
+function runphysics(e::Engine)
+    p = e.phys
+    ui = e.ui
+    resettime!(p) # reset sim and world clocks to 0
 
-    return nothing
-end
+    try
+        while true
+            shouldexit, lastrender, reversed, paused, refrate, = @lock_nofail ui.lock begin
+                ui.shouldexit, ui.lastrender, ui.reversed, ui.paused, ui.refreshrate
+            end
 
-"""
-    close_viewer!(e::Engine)
+            if shouldexit
+                break
+            elseif (time() - lastrender) > 1 / e.min_refreshrate
+                yield()
+                continue
+            else
+                @lock p.lock begin
+                    elapsedworld = time(p.timer)
 
-Close the viewer when done.
-"""
-function close_viewer!(e::Engine)
-    GLFW.DestroyWindow(e.manager.state.window)
+                    # advance sim
+                    if ui.paused
+                        pausestep!(p, mode(e))
+                    elseif ui.reversed && p.elapsedsim > elapsedworld
+                        reversestep!(p, mode(e))
+                        p.elapsedsim -= timestep(p.model)
+                    elseif !ui.reversed && p.elapsedsim < elapsedworld
+                        forwardstep!(p, mode(e))
+                        p.elapsedsim += timestep(p.model)
+                    end
+                end
+            end
+        end
+    finally
+        @lock ui.lock begin
+            ui.shouldexit = true
+        end
+    end
     return nothing
 end
